requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pydantic
google-generativeai
fastapi
uvicorn[standard]
itsdangerous

